refactor: beverage 관련 .java 파일 이동
