+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/item/domain/beverage/BeverageItem.java
@@ -1,6 +1,10 @@
-package git_kkalnane.backend.starbucks.item.domain;
+package git_kkalnane.backend.starbucks.item.domain.beverage;
 
 import git_kkalnane.backend.starbucks._global.entity.BaseTimeEntity;
+import git_kkalnane.backend.starbucks.item.domain.ItemStatus;
+import git_kkalnane.backend.starbucks.item.domain.beverage.enums.BeverageShotOption;
+import git_kkalnane.backend.starbucks.item.domain.beverage.enums.BeverageSizeOption;
+import git_kkalnane.backend.starbucks.item.domain.beverage.enums.BeverageTemperatureOption;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
